LayoutTests/platform/chromium-win/fast/css/getComputedStyle/getComputedStyle-margin-percentage-expected.txt

Test calling getPropertyValue on computed margin styles.

margin-left: 258.7166748046875px
margin-top: 78.4000015258789px
margin-right: 113px
margin-bottom: 0px
Test
